requirements.txt
appdirs==1.4.4
benzinga==1.21
certifi==2020.12.5
chardet==4.0.0
distlib==0.3.1
filelock==3.0.12
idna==2.10
inflection==0.5.1
lxml==4.6.3
more-itertools==8.7.0
multitasking==0.0.9
numpy==1.20.2
pandas==1.2.4
python-dateutil==2.8.1
python-dotenv==0.17.0
pytz==2021.1
Quandl==3.6.1
requests==2.25.1
six==1.15.0
structlog==21.1.0
urllib3==1.25.10
virtualenv==20.4.4
yfinance==0.1.59
